+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- kv: `getLong` and `putLong` methods;
 - concord-console: "download state" button to the process status
 page;  
 - concord-agent, concord-server: detect orphaned or stalled